Improve stories. Add onBefore callbacks
