+++ .prettierignore
@@ -4,4 +4,6 @@ dist/
 coverage/
 src/**/*/types.js
 src/**/*/*.types.js
+src/styles/configs/_colorway.scss
+src/styles/configs/colorway.ts
 .remake
